se elimino el import del componente cocina en el app.jsx y se agregó en el login el navigate si el usuario no está activo

+++ src/components/Login/Login.jsx
@@ -7,7 +7,7 @@ import { useForm } from "react-hook-form";
 import Swal from 'sweetalert2'
 import "./login.css";
 
-const Login = ({ setUsuarioLogueado, usuarioLogueado }) => {
+const Login = ({ setUsuarioLogueado }) => {
   const {
     register,
     handleSubmit,
@@ -19,9 +19,23 @@ const Login = ({ setUsuarioLogueado, usuarioLogueado }) => {
   const login = async (usuario) => {
     try {
       const respuesta = await loginUsuario(usuario);
-      if (respuesta.status === 200) {
-        const datos = await respuesta.json();
+      const datos = await respuesta.json();
+      console.log(datos)
+      if(datos.estado === 'inactivo') {
+        Swal.fire({
+          position: "center",
+          icon: "info",
+          title: "¡Lo Sentimos!",
+          text: "Su cuenta se encuentra momentáneamente inactiva",
+          showConfirmButton: false,
+          timer: 1500
+        });
+        
+        navegacion("/ingresar");
+        return
+      }
 
+      if (respuesta.status === 200) {
         sessionStorage.setItem(
           "usuarioLotus",
           JSON.stringify({ email: datos.email, token: datos.token, tipoUsuario: datos.tipoUsuario })
@@ -43,8 +57,6 @@ const Login = ({ setUsuarioLogueado, usuarioLogueado }) => {
         }
 
       } else {
-        Swal.fire("Ocurrió un error", "Correo o contraseña incorrectos", "error");
-
         Swal.fire({
           position: "center",
           icon: "error",
